Convert Timelast Casts to TS and make it reusable by switching props
